passport setup
Set up the express session in app.js and also created the passport.js
for setting up the local authentication strategy, added the output of
error message in the strategy and also added redirecting to login page
if authentication fails.

Added some variables to res.locals so that the logged in username and
log in status can be accessed by all the rendering actions.